caidao's exploit module
